saga_api, saga_gui, tools: fix variable declarations of "for loop" counters (sLong and size_t datatypes); includes also some other fixes (e.g. sLong for Get_NCells() loops, format specifiers, and other changes related to the "sLong number of records" changes)

+++ saga-gis/src/tools/grid/grid_analysis/Grid_IMCORR.cpp
@@ -1128,18 +1128,18 @@ void CGrid_IMCORR::gnorm(std::vector<double>& CCNORM , std::vector<double>& pkva
 	std::vector<double>ser, ref;
 	//create arrays
 	ser.push_back(0.0); // dummy index to allow fortran conform indexing 
-	for(int i = 0; i<ChipSearch.size(); i++)
+	for(size_t i=0; i<ChipSearch.size(); i++)
 	{
-		for(int ii = 0; ii<ChipSearch[0].size(); ii++)
+		for(size_t ii=0; ii<ChipSearch[0].size(); ii++)
 		{
 			ser.push_back(ChipSearch[i][ii]);
 		}
 	}
 
 	ref.push_back(0.0); // dummy index to allow fortran conform indexing 
-	for(int i = 0; i<ChipRef.size(); i++)
+	for(size_t i=0; i<ChipRef.size(); i++)
 	{
-		for(int ii = 0; ii<ChipRef[0].size(); ii++)
+		for(size_t ii=0; ii<ChipRef[0].size(); ii++)
 		{
 			ref.push_back(ChipRef[i][ii]);
 		}
@@ -1353,13 +1353,13 @@ void CGrid_IMCORR::cross(std::vector<double>& UNORMC , std::vector<std::vector<d
 
 	std::vector<std::vector<double> > ChipRef2;
 	ChipRef2.resize(ChipSearch[0].size());
-	for(int i = 0; i < ChipSearch[0].size(); i++)
+	for(size_t i=0; i<ChipSearch[0].size(); i++)
 		ChipRef2[i].resize(ChipSearch.size(),0.0);
 
 	// zero extent chipref to search chip size
-	for(int i = 0; i<ChipRef.size(); i++)
+	for(size_t i=0; i<ChipRef.size(); i++)
 	{
-		for(int ii = 0; ii<ChipRef[0].size(); ii++)
+		for(size_t ii=0; ii<ChipRef[0].size(); ii++)
 		{
 			ChipRef2[i][ii] = ChipRef[i][ii];
 		}
@@ -1369,18 +1369,18 @@ void CGrid_IMCORR::cross(std::vector<double>& UNORMC , std::vector<std::vector<d
 	std::vector<double>ser, ref;
 	//create arrays
 	ser.push_back(0.0); // dummy index to allow fortran conform indexing 
-	for(int i = 0; i<ChipSearch.size(); i++)
+	for(size_t i=0; i<ChipSearch.size(); i++)
 	{
-		for(int ii = 0; ii<ChipSearch[0].size(); ii++)
+		for(size_t ii=0; ii<ChipSearch[0].size(); ii++)
 		{
 			ser.push_back(ChipSearch[i][ii]);
 		}
 	}
 
 	ref.push_back(0.0); // dummy index to allow fortran conform indexing 
-	for(int i = 0; i<ChipRef2.size(); i++)
+	for(size_t i=0; i<ChipRef2.size(); i++)
 	{
-		for(int ii = 0; ii<ChipRef2[0].size(); ii++)
+		for(size_t ii=0; ii<ChipRef2[0].size(); ii++)
 		{
 			ref.push_back(ChipRef2[i][ii]);
 		}
@@ -1395,7 +1395,7 @@ void CGrid_IMCORR::cross(std::vector<double>& UNORMC , std::vector<std::vector<d
 	std::vector<double>cser, cref;
 	cser.push_back(0.0);
 	cref.push_back(0.0);
-	for (int i = 1; i < ser.size(); i++)
+	for (size_t i=1; i<ser.size(); i++)
 	{
 		cser.push_back(ser[i]);
 		cser.push_back(0.0);
@@ -1408,7 +1408,7 @@ void CGrid_IMCORR::cross(std::vector<double>& UNORMC , std::vector<std::vector<d
 	fft2(cref,nsnew,1);
 
 	// make point by multiplication of ft of search ft with conjugate of reference image
-	for(int i = 1; i<cser.size(); i+=2)
+	for(size_t i=1; i<cser.size(); i+=2)
 	{
 		// cser[i] = cser[i] * conjg(cref[i])
 		double temp_cser_real = cser[i];
@@ -1554,7 +1554,7 @@ void CGrid_IMCORR::fft2(std::vector<double>& data, std::vector<int>nel, int isig
 int CGrid_IMCORR::decimal(std::vector<int> Bin) 
 {
 	int Sum = 0;
-	for (int i = 0; i < Bin.size(); i++)
+	for (size_t i=0; i<Bin.size(); i++)
 	{	
 		int Exponent = (int)(Bin.size()-1)-i;
 		Sum+= Bin[i] * (int)pow(2.0, Exponent);
